Submission filtering and searching updated.
Removed multiple PRAW requests as the user search results were often cached and inaccurate. Now the "new" submisisons are cached and searched internally. It only needs 1 request per cycle and is much more reliable.

Also on startup, it simply reports the latest post instead of using a time. The time can still be specified if you want it to report the old way.